refactor: unify path detection logic into pure utility
Extract resolveHomeDir and resolveConfigDir from diagnostics.js, config.js,
and analyze.js into a single shared utility module (src/utils/paths.js).
This removes duplicate implementations, guarantees consistent OS-specific
path resolution across the app, and avoids triggering env-var constraints
when standalone scripts import the logic.
