+++ src/config.js
@@ -15,13 +15,11 @@ if (!SECRET_KEY) {
 const TUNNEL_TOKEN = process.env.TUNNEL_TOKEN;
 
 // --- Paths ---
-// [P3 Fix] Resolve home dir using OPENCLAW_HOME → HOME → os.homedir()
-// aligned with OpenClaw's src/infra/home-dir.ts
-const HOME_DIR = process.env.OPENCLAW_HOME
-    || process.env.HOME
-    || process.env.USERPROFILE
-    || os.homedir();
-const STATE_DIR = process.env.OPENCLAW_STATE_DIR || path.join(HOME_DIR, '.openclaw');
+// Shared home/config directory resolution, aligned with OpenClaw's src/infra/home-dir.ts.
+const { resolveHomeDir, resolveConfigDir } = require('./utils/paths');
+
+const HOME_DIR = resolveHomeDir();
+const STATE_DIR = resolveConfigDir();
 const APP_DIR = path.resolve(__dirname, '..');
 
 const LOG_DIR = path.join(APP_DIR, 'data/logs');
@@ -44,4 +42,6 @@ module.exports = {
     ID_FILE,
     ANALYZE_SCRIPT,
     CACHE_TTL_MS,
+    resolveHomeDir,
+    resolveConfigDir,
 };
